Start next and the proxy server programatically.
Previously, all the servers (api/next/proxy) would be started
at the same time. Any requests to the proxy server before next
was ready would cause everything to crash. This was "fixed" by
adding a listen to the proxy 'error' event, but didn't solve the
race condition.

This patch rebuilds the proxy server to wait for the next server
to successfully start, then start listening itself. Also the api
server is eliminated, with it becoming an express sub app.